[MIPS] Disable block linking for partial flushing and make sure reg_a0 has original write address before flush call
I will preface this with I don't know much about MIPS assembly, so this may be wrong!!!

reg_a0 (MIPS reg $4 i.e. the first argument for a C function call) needs to be the full original address prior to the partial flush call.  Had to move some code around to achieve this.  a) it now may not work and b) I may have got some of the code wrong with my assumptions about how MIPS instructions and registers work.